refactor: extract catalyst and unity-runner as adapter components
Continues the WKC alignment by extracting the two clearest I/O
boundaries in the conversion code into adapter components.

Decisions on what NOT to refactor (recorded for the PR description):
- asset-reuse.ts: 1027 lines, deeply intertwined with ~30 tests that
  import its module-level functions directly; the only stateful piece
  (probeHitCache LRU) is already managed explicitly by tests.
  Component conversion is a separate PR's worth of work.
- conversion-task.ts: stays as a logic module — the orchestrator
  already wraps it as a component. Wrapping its three async functions
  (executeConversion, executeLODConversion, executeTriagePass) into
  yet another component layer would add ceremony without benefit.
- gltf-deps.ts, extensions.ts, s3-helpers.ts, timer.ts: pure utility
  functions (no I/O, no state). WKC says components are stateful or
  do I/O — these don't qualify.
- has-content-changed-task.ts: legacy fallback marked for removal
  once the per-asset reuse path proves in production.
- ensure-ulf.ts: one-shot startup-time license file decode.

src/adapters/catalyst/ (new):
  - HTTP client for Decentraland catalysts and worlds-content-server.
  - getActiveEntity uses GET /contents/{CID} (works on both backends);
    optional AbortController-based timeout for callers that need a
    bound (e.g., the migration script).
  - getEntities uses POST /entities/active for pointer lookups.
  - Standalone fetchActiveEntity also exported for CLI scripts that
    don't construct the full components container — same logic, two
    surfaces, byte-identical behavior.

src/adapters/unity-runner/ (new):
  - Encapsulates the Unity child-process spawn (runConversion +
    runLodsConversion) plus the previously-separate run-command.ts
    helper.
  - Owns the unity-runner logger and metric emission; methods take
    only per-invocation options.
  - CLI argument validation (joinValidatedHashes for CIDs) stays
    inside the component.

Wired into components.ts and types.ts; orchestrator's Components
Pick widened to include catalyst + unityRunner. conversion-task.ts
calls them via components.catalyst.* / components.unityRunner.*
instead of importing module-level functions. has-content-changed-task
takes catalyst as a parameter.

CLI scripts (migrate-to-canonical.ts, test-conversion.ts) and
integration tests updated:
  - migrate uses fetchActiveEntity exported from the catalyst adapter
  - test-conversion constructs the catalyst + unity-runner components
  - integration tests inject mock components via the components
    object instead of jest.mock-ing the (now-deleted) modules
  - run-conversion.spec.ts unit test rewritten to drive the
    unity-runner component; standalone startManifestBuilder describe
    block removed (its coverage is preserved by the existing doISS
    integration tests)

372 tests pass.
