+++ consumer-server/src/adapters/catalyst/index.ts
@@ -1,0 +1,2 @@
+export { createCatalystComponent, fetchActiveEntity } from './component'
+export type { ICatalystComponent } from './types'
