+++ consumer-server/test/integration/execute-conversion.spec.ts
@@ -21,26 +21,18 @@ import {
 } from '../../src/logic/asset-reuse'
 import { buildGlb } from '../helpers/glb-fixtures'
 
-jest.mock('../../src/logic/run-conversion', () => ({
-  runConversion: jest.fn(),
-  runLodsConversion: jest.fn()
-}))
-jest.mock('../../src/logic/fetch-entity-by-pointer', () => ({
-  getActiveEntity: jest.fn(),
-  getEntities: jest.fn()
-}))
 jest.mock('../../src/logic/has-content-changed-task', () => {
   const real = jest.requireActual('../../src/logic/has-content-changed-task')
   return { ...real, hasContentChange: jest.fn(async () => true) }
 })
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const MockAws = require('mock-aws-s3')
-import { runConversion } from '../../src/logic/run-conversion'
-import { getActiveEntity } from '../../src/logic/fetch-entity-by-pointer'
 import { executeConversion } from '../../src/logic/conversion-task'
 
-const mockedRunConversion = runConversion as jest.Mock
-const mockedGetActiveEntity = getActiveEntity as jest.Mock
+const mockedRunConversion = jest.fn()
+const mockedRunLodsConversion = jest.fn()
+const mockedGetActiveEntity = jest.fn()
+const mockedGetEntities = jest.fn()
 const originalNativeFetch = globalThis.fetch
 let mockedFetch: jest.Mock
 
@@ -71,7 +63,21 @@ function buildComponents(bucketBasePath: string, params: Params = {}) {
     captureException: jest.fn()
   } as any
 
-  return { config, cdnS3, sentry }
+  // Component-style mocks: replace the previous jest.mock('../../src/logic/...')
+  // pattern with explicit injection through the components object. Tests then
+  // exercise executeConversion against the same mock surface (mockedRunConversion,
+  // mockedGetActiveEntity, mockedGetEntities) but routed via dependency
+  // injection — which is what the orchestrator does in production too.
+  const catalyst = {
+    getActiveEntity: mockedGetActiveEntity,
+    getEntities: mockedGetEntities
+  }
+  const unityRunner = {
+    runConversion: mockedRunConversion,
+    runLodsConversion: mockedRunLodsConversion
+  }
+
+  return { config, cdnS3, sentry, catalyst, unityRunner }
 }
 
 // Wire up native fetch so glb/gltf digest reads return declared fixtures and
@@ -348,7 +354,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
       // Mocked Unity: writes the non-cached bundles to outDirectory and returns 0.
       // GLB output uses the composite filename (Unity-side depsDigest naming);
       // BIN stays at the bare `{hash}_{target}` form.
-      mockedRunConversion.mockImplementation(async (_logger: any, _components: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         expect(options.depsDigestByHash.get('hGlb')).toBe(perGlbDigest)
         expect(options.depsDigestByHash.get('hNewGlb')).toBe(perGlbDigest)
         await fs.mkdir(options.outDirectory, { recursive: true })
@@ -371,7 +377,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
       expect(mockedRunConversion).toHaveBeenCalledTimes(1)
 
       // Assert -cachedHashes contained only the GLB.
-      const passedOptions = mockedRunConversion.mock.calls[0][2]
+      const passedOptions = mockedRunConversion.mock.calls[0][0]
       expect((passedOptions.cachedHashes ?? []).sort()).toEqual(['hGlb'])
 
       // New bundles landed at their canonical paths.
@@ -414,7 +420,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, 'hOnlyGlb_windows'), 'freshly-converted')
         return 0
@@ -440,7 +446,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should NOT pass a cachedHashes list to Unity', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should upload the freshly-converted bundle to the entity-scoped path (reuse path is gated off by the kill switch)', async () => {
@@ -480,7 +486,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, composedGlbFilename), 'fresh-glb')
         return 0
@@ -506,7 +512,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should not pass cachedHashes to Unity (force skips the cache probe)', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should overwrite existing canonical bytes with the freshly converted output', async () => {
@@ -556,7 +562,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, cachedGlbFilename), 'fresh-cached')
         await fs.writeFile(path.join(options.outDirectory, missingGlbFilename), 'fresh-missing')
@@ -579,7 +585,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should not pass cachedHashes to Unity even though one asset would have matched', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should overwrite the pre-existing canonical bytes for the would-have-been-cached asset', async () => {
@@ -610,7 +616,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, 'hGlb_windows'), 'freshly-converted')
         return 0
@@ -636,7 +642,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should NOT pass a cachedHashes list to Unity', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should upload the freshly-converted bundle to the entity-scoped path (reuse path is gated off by doISS)', async () => {
@@ -665,7 +671,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, 'hWearableGlb_windows'), 'new-bundle')
         return 0
@@ -691,7 +697,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should NOT pass a cachedHashes list to Unity', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should upload the bundle to the entity-scoped path', async () => {
@@ -806,7 +812,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         return realHead(params)
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         // Probe failure must NOT zero out depsDigestByHash — otherwise glb
         // bundles would revert to bare names and reintroduce the cross-scene
         // collision.
@@ -833,7 +839,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should pass no cached hashes to Unity (cacheResult was set to null)', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should still upload to the canonical prefix at the composite glb path', async () => {
@@ -873,7 +879,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: { main: 'index.js' }
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         runConversionOptions = options
         // Unity ignores the broken glb (per `-skippedHashes`) and produces
         // a bundle for the texture, which the integration test asserts
@@ -1005,7 +1011,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     beforeEach(async () => {
       mockedGetActiveEntity.mockRejectedValue(new Error('catalyst unavailable'))
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, 'hNoEntity_windows'), 'bundle')
         return 0
@@ -1031,7 +1037,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
     })
 
     it('should NOT pass a cachedHashes list to Unity (reuse path requires the entity)', () => {
-      expect(mockedRunConversion.mock.calls[0][2].cachedHashes).toBeUndefined()
+      expect(mockedRunConversion.mock.calls[0][0].cachedHashes).toBeUndefined()
     })
 
     it('should upload the bundle to the entity-scoped path', async () => {
@@ -1056,7 +1062,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         content: [{ file: 'model.glb', hash: 'hGlb' }],
         metadata: {}
       })
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, 'out'), 'x')
         return 0
@@ -1135,7 +1141,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(options.outDirectory, { recursive: true })
         await fs.writeFile(path.join(options.outDirectory, 'hGlb_windows'), 'bundle')
         return 0
@@ -1178,7 +1184,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         content: [{ file: 'model.glb', hash: 'hGlb' }],
         metadata: {}
       })
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.writeFile(options.logFile, 'unity log')
         await fs.mkdir(options.outDirectory, { recursive: true })
@@ -1223,7 +1229,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         content: [{ file: 'model.glb', hash: 'hGlb' }],
         metadata: {}
       })
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.writeFile(options.logFile, 'unity log')
         await fs.mkdir(options.outDirectory, { recursive: true })
@@ -1297,7 +1303,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         content: [{ file: 'model.glb', hash: 'hGlb' }],
         metadata: {}
       })
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.writeFile(options.logFile, 'unity log')
         await fs.mkdir(options.outDirectory, { recursive: true })
@@ -1502,7 +1508,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
       // mock must create the logFile before throwing, because the outer
       // catch in conversion-task.ts reads it first — if it doesn't exist,
       // ENOENT replaces the original error and masks the test's assertion.
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.writeFile(options.logFile, 'partial log')
         throw new Error('simulated Unity crash')
@@ -1580,7 +1586,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.writeFile(options.logFile, 'unity log')
         await fs.mkdir(options.outDirectory, { recursive: true })
@@ -1645,7 +1651,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         content: [{ file: 'model.glb', hash: 'hGlb' }],
         metadata: {}
       })
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.writeFile(options.logFile, 'unity log')
         await fs.mkdir(options.outDirectory, { recursive: true })
@@ -1712,7 +1718,7 @@ describe('when executing a conversion with asset-reuse enabled', () => {
         metadata: {}
       })
 
-      mockedRunConversion.mockImplementation(async (_l: any, _c: any, options: any) => {
+      mockedRunConversion.mockImplementation(async (options: any) => {
         await fs.mkdir(path.dirname(options.logFile), { recursive: true })
         await fs.mkdir(options.outDirectory, { recursive: true })
         // Write the expected log file that conversion-task's finally tries to upload.
