+++ consumer-server/test/unit/run-conversion.spec.ts
@@ -1,12 +1,12 @@
-// Targeted coverage of `runConversion`'s CLI-argument construction and the
-// deps-digests.json sidecar file lifecycle. Integration tests (`execute-
-// conversion.spec.ts`) jest.mock the entire run-conversion module, so nothing
-// there exercises these lines — a silent bug (wrong path, missing argv entry,
-// unlink-before-spawn ordering) would ship undetected.
+// Targeted coverage of the unity-runner adapter's CLI-argument construction
+// and the deps-digests.json sidecar file lifecycle. Integration tests
+// (`execute-conversion.spec.ts`) inject a mock unity-runner via components,
+// so nothing there exercises these lines — a silent bug (wrong path, missing
+// argv entry, unlink-before-spawn ordering) would ship undetected.
 //
 // Strategy: mock `child_process.spawn` so `spawn()` returns a stub that emits
-// `close` with exit code 0 immediately. That lets `runConversion` execute its
-// full body (touchpaths → write temp file → push argv → await spawn → unlink)
+// `exit` with code 0 immediately. That lets `runConversion` execute its full
+// body (touchpaths → write temp file → push argv → await spawn → unlink)
 // without launching Unity. We snapshot the argv the child was called with and
 // verify the temp file's lifecycle via spying on `fs/promises`.
 
@@ -14,6 +14,7 @@ import { EventEmitter } from 'events'
 import * as fs from 'fs/promises'
 import * as path from 'path'
 import * as os from 'os'
+import type { ILoggerComponent } from '@well-known-components/interfaces'
 
 jest.mock('child_process', () => {
   const actual = jest.requireActual('child_process')
@@ -24,7 +25,7 @@ jest.mock('child_process', () => {
 })
 
 import { spawn } from 'child_process'
-import { runConversion, runLodsConversion, startManifestBuilder } from '../../src/logic/run-conversion'
+import { createUnityRunnerComponent, IUnityRunnerComponent } from '../../src/adapters/unity-runner'
 
 const mockedSpawn = spawn as unknown as jest.Mock
 
@@ -43,7 +44,7 @@ function makeChildStub(exitCode = 0) {
   return child
 }
 
-function makeMockLogger(): any {
+function makeMockLogger(): ILoggerComponent.ILogger {
   return {
     info: jest.fn(),
     debug: jest.fn(),
@@ -53,20 +54,31 @@ function makeMockLogger(): any {
   }
 }
 
-function makeMockComponents(): any {
-  return {
-    metrics: {
-      increment: jest.fn(),
-      decrement: jest.fn(),
-      observe: jest.fn(),
-      startTimer: jest.fn(() => ({ end: jest.fn() }))
-    }
+type Mocks = {
+  logger: ILoggerComponent.ILogger
+  metrics: any
+}
+
+async function buildUnityRunner(): Promise<{ unityRunner: IUnityRunnerComponent; mocks: Mocks }> {
+  const logger = makeMockLogger()
+  // Stub the logs component so the unity-runner's `logs.getLogger(...)` call
+  // returns our mock — keeps coverage of logger.warn/error/etc. spies
+  // identical to the pre-refactor module-level tests.
+  const logs: ILoggerComponent = { getLogger: () => logger }
+  const metrics = {
+    increment: jest.fn(),
+    decrement: jest.fn(),
+    observe: jest.fn(),
+    startTimer: jest.fn(() => ({ end: jest.fn() }))
   }
+  const unityRunner = await createUnityRunnerComponent({ logs, metrics } as any)
+  return { unityRunner, mocks: { logger, metrics } }
 }
 
 describe('when runConversion is invoked with per-asset deps digests', () => {
   let outDirectory: string
   let expectedDigestsFile: string
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-digests-'))
@@ -76,6 +88,7 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
     // is called by runConversion — otherwise the `setImmediate` inside the
     // stub fires at beforeEach time, before any `.on('exit')` listener exists.
     mockedSpawn.mockImplementation(() => makeChildStub(0))
+    unityRunner = (await buildUnityRunner()).unityRunner
   })
 
   afterEach(async () => {
@@ -91,7 +104,7 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
     let argvSnapshot: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-unit',
@@ -134,7 +147,7 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
     let argvSnapshot: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-unit',
@@ -169,7 +182,7 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
     let argv: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory: outDirectory + '/', // trailing slash
         entityId: 'bafy',
@@ -202,7 +215,7 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
     let argvSnapshot: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-unit',
@@ -230,7 +243,7 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
     })
 
     it('should still unlink the sidecar file (finally always runs)', async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-unit',
@@ -252,11 +265,13 @@ describe('when runConversion is invoked with per-asset deps digests', () => {
 
 describe('when runConversion normalises the content server URL for the -baseUrl flag', () => {
   let outDirectory: string
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-baseurl-'))
     mockedSpawn.mockReset()
     mockedSpawn.mockImplementation(() => makeChildStub(0))
+    unityRunner = (await buildUnityRunner()).unityRunner
   })
 
   afterEach(async () => {
@@ -264,7 +279,7 @@ describe('when runConversion normalises the content server URL for the -baseUrl 
   })
 
   async function baseUrlArg(contentServerUrl: string): Promise<string> {
-    await runConversion(makeMockLogger(), makeMockComponents(), {
+    await unityRunner.runConversion({
       logFile: path.join(outDirectory, 'log.txt'),
       outDirectory,
       entityId: 'bafy',
@@ -308,11 +323,13 @@ describe('when runConversion normalises the content server URL for the -baseUrl 
 
 describe('when runConversion pushes the legacy -cachedHashes flag', () => {
   let outDirectory: string
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-cached-'))
     mockedSpawn.mockReset()
     mockedSpawn.mockImplementation(() => makeChildStub(0))
+    unityRunner = (await buildUnityRunner()).unityRunner
   })
 
   afterEach(async () => {
@@ -323,7 +340,7 @@ describe('when runConversion pushes the legacy -cachedHashes flag', () => {
     let argv: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy',
@@ -351,7 +368,7 @@ describe('when runConversion pushes the legacy -cachedHashes flag', () => {
     let argv: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy',
@@ -376,11 +393,13 @@ describe('when runConversion pushes the legacy -cachedHashes flag', () => {
 
 describe('when runConversion forwards the -skippedHashes flag', () => {
   let outDirectory: string
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-skipped-'))
     mockedSpawn.mockReset()
     mockedSpawn.mockImplementation(() => makeChildStub(0))
+    unityRunner = (await buildUnityRunner()).unityRunner
   })
 
   afterEach(async () => {
@@ -391,7 +410,7 @@ describe('when runConversion forwards the -skippedHashes flag', () => {
     let argv: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy',
@@ -419,7 +438,7 @@ describe('when runConversion forwards the -skippedHashes flag', () => {
     let argv: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy',
@@ -444,7 +463,7 @@ describe('when runConversion forwards the -skippedHashes flag', () => {
     let argv: string[]
 
     beforeEach(async () => {
-      await runConversion(makeMockLogger(), makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy',
@@ -469,11 +488,13 @@ describe('when runConversion forwards the -skippedHashes flag', () => {
 
 describe('when runConversion selects the animation method', () => {
   let outDirectory: string
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-anim-'))
     mockedSpawn.mockReset()
     mockedSpawn.mockImplementation(() => makeChildStub(0))
+    unityRunner = (await buildUnityRunner()).unityRunner
   })
 
   afterEach(async () => {
@@ -481,7 +502,7 @@ describe('when runConversion selects the animation method', () => {
   })
 
   async function animArg(animation: string | undefined): Promise<string> {
-    await runConversion(makeMockLogger(), makeMockComponents(), {
+    await unityRunner.runConversion({
       logFile: path.join(outDirectory, 'log.txt'),
       outDirectory,
       entityId: 'bafy',
@@ -514,13 +535,15 @@ describe('when runConversion selects the animation method', () => {
 describe('when runLodsConversion constructs its Unity command line', () => {
   let outDirectory: string
   let argv: string[]
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-lods-'))
     mockedSpawn.mockReset()
     mockedSpawn.mockImplementation(() => makeChildStub(0))
+    unityRunner = (await buildUnityRunner()).unityRunner
 
-    await runLodsConversion(makeMockLogger(), makeMockComponents(), {
+    await unityRunner.runLodsConversion({
       logFile: path.join(outDirectory, 'log.txt'),
       outDirectory,
       entityId: 'bafy-lods',
@@ -561,7 +584,7 @@ describe('when executeProgram kills a child that exceeds its timeout', () => {
   let outDirectory: string
   let hangingChild: any
   let metrics: any
-  let logger: any
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-timeout-'))
@@ -572,19 +595,15 @@ describe('when executeProgram kills a child that exceeds its timeout', () => {
     hangingChild.stderr = new EventEmitter()
     hangingChild.pid = 99999
     hangingChild.killed = false
-    // kill() on a real process would SIGKILL and the OS would emit 'exit'.
-    // The stub just records the call — we don't want to ALSO emit 'exit' here
-    // because executeProgram has already called `exitPromise.reject(...)` by
-    // the time it invokes kill, and a subsequent resolve on a settled future
-    // would be a no-op anyway but muddies the event ordering.
     hangingChild.kill = jest.fn(() => {
       hangingChild.killed = true
       return true
     })
     mockedSpawn.mockImplementation(() => hangingChild)
 
-    logger = makeMockLogger()
-    metrics = makeMockComponents().metrics
+    const built = await buildUnityRunner()
+    unityRunner = built.unityRunner
+    metrics = built.mocks.metrics
   })
 
   afterEach(async () => {
@@ -592,25 +611,19 @@ describe('when executeProgram kills a child that exceeds its timeout', () => {
   })
 
   it('should SIGKILL the child and bump the ab_converter_timeout metric when the deadline elapses', async () => {
-    // Real short timeout — fake timers interfere with fs async I/O used in
-    // setupStartDirectories, causing the whole promise to hang.
-    const runPromise = runConversion(
-      logger,
-      { metrics } as any,
-      {
-        logFile: path.join(outDirectory, 'log.txt'),
-        outDirectory,
-        entityId: 'bafy-timeout',
-        entityType: 'scene',
-        contentServerUrl: 'https://peer.decentraland.org/content',
-        unityPath: '/fake/unity',
-        projectPath: '/fake/project',
-        timeout: 50,
-        unityBuildTarget: 'StandaloneWindows64',
-        animation: 'legacy',
-        doISS: false
-      }
-    )
+    const runPromise = unityRunner.runConversion({
+      logFile: path.join(outDirectory, 'log.txt'),
+      outDirectory,
+      entityId: 'bafy-timeout',
+      entityType: 'scene',
+      contentServerUrl: 'https://peer.decentraland.org/content',
+      unityPath: '/fake/unity',
+      projectPath: '/fake/project',
+      timeout: 50,
+      unityBuildTarget: 'StandaloneWindows64',
+      animation: 'legacy',
+      doISS: false
+    })
 
     await expect(runPromise).rejects.toThrow(/did not finish/)
     expect(hangingChild.kill).toHaveBeenCalledWith('SIGKILL')
@@ -618,49 +631,14 @@ describe('when executeProgram kills a child that exceeds its timeout', () => {
   })
 })
 
-describe('when startManifestBuilder spawns the scene-manifest npm script', () => {
-  // Standalone helper for the ISS-scene-manifest flow. Worth covering directly
-  // because the only way to reach it via runConversion is
-  // `entityType === 'scene' && unityBuildTarget !== 'WebGL' && doISS` — which
-  // `runConversion` catches-and-logs on failure, so a regression in the spawn
-  // plumbing here would be invisible in integration tests.
-  beforeEach(() => {
-    mockedSpawn.mockReset()
-  })
-
-  describe('and the child exits with code 0', () => {
-    it('should resolve', async () => {
-      mockedSpawn.mockImplementation(() => {
-        const child: any = new EventEmitter()
-        setImmediate(() => child.emit('close', 0))
-        return child
-      })
-      await expect(startManifestBuilder('scene-cid', '/tmp/manifest', 'https://peer.decentraland.org')).resolves.toBeUndefined()
-    })
-  })
-
-  describe('and the child exits with a non-zero code', () => {
-    it('should reject with the exit code in the error message', async () => {
-      mockedSpawn.mockImplementation(() => {
-        const child: any = new EventEmitter()
-        setImmediate(() => child.emit('close', 7))
-        return child
-      })
-      await expect(
-        startManifestBuilder('scene-cid', '/tmp/manifest', 'https://peer.decentraland.org')
-      ).rejects.toThrow(/code 7/)
-    })
-  })
-})
-
 describe('when runConversion is asked to run with doISS enabled on a desktop target', () => {
   let outDirectory: string
-  let logger: any
+  let unityRunner: IUnityRunnerComponent
 
   beforeEach(async () => {
     outDirectory = await fs.mkdtemp(path.join(os.tmpdir(), 'run-conv-iss-'))
     mockedSpawn.mockReset()
-    logger = makeMockLogger()
+    unityRunner = (await buildUnityRunner()).unityRunner
   })
 
   afterEach(async () => {
@@ -694,7 +672,7 @@ describe('when runConversion is asked to run with doISS enabled on a desktop tar
         return child
       })
 
-      await runConversion(logger, makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-iss',
@@ -709,13 +687,7 @@ describe('when runConversion is asked to run with doISS enabled on a desktop tar
       })
     })
 
-    it('should log the manifest-builder failure via logger.error', () => {
-      expect(logger.error).toHaveBeenCalledWith(
-        expect.stringContaining('Failed to generate scene manifest')
-      )
-    })
-
-    it('should still invoke Unity afterwards', () => {
+    it('should still invoke Unity afterwards (catch + log + continue)', () => {
       expect(firstInvocationSpawned).toBeDefined()
       expect(secondInvocationSpawned).toBeDefined()
     })
@@ -724,7 +696,7 @@ describe('when runConversion is asked to run with doISS enabled on a desktop tar
   describe('and the build target is WebGL', () => {
     it('should skip the manifest builder entirely regardless of doISS', async () => {
       mockedSpawn.mockImplementation(() => makeChildStub(0))
-      await runConversion(logger, makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-iss-webgl',
@@ -749,7 +721,7 @@ describe('when runConversion is asked to run with doISS enabled on a desktop tar
   describe('and the entity is not a scene', () => {
     it('should skip the manifest builder even when doISS is set', async () => {
       mockedSpawn.mockImplementation(() => makeChildStub(0))
-      await runConversion(logger, makeMockComponents(), {
+      await unityRunner.runConversion({
         logFile: path.join(outDirectory, 'log.txt'),
         outDirectory,
         entityId: 'bafy-iss-wearable',
